Revert "Types"
This reverts commit 736a2582ee2b4104e6556e2fa56849f23ee7fe60.

+++ include/seqan/kmer/kmer_ibf.h
@@ -172,7 +172,7 @@ public:
      * \param threads Number of threads to use.
      */
     template<typename TInt>
-    void clear(std::vector<uint32_t> const & bins, TInt&& threads)
+    void clear(std::vector<uint16_t> const & bins, TInt&& threads)
     {
         std::vector<std::future<void>> tasks;
         uint64_t chunkBlocks = filterVector.chunkSize / filterVector.blockBitSize;
